scoreboard와 namecard를 유저의 id로 변경

+++ Server/FreezeBombServer/FreezeBombServer/Server.h
@@ -114,8 +114,8 @@ private:
 	CGameTimer gameTimer;
 	CHeightMapImage* heightMap;
 	XMFLOAT3 gravity;
-	float roundStartTime;
-	float roundCurrTime;
+	unsigned short roundStartTime;
+	unsigned short roundCurrTime;
 	int clientCount;
 	int readyCount;
 	int hostId;
